refactor: split tx.rs (2232 lines) into 4 submodules
Split the monolithic transaction engine into focused submodules:

- tx/validate.rs (135 lines): op_label, validate_operation,
  validate_plan_operations
- tx/output.rs (252 lines): TxOutput, TxChange, TxSearchResult,
  TxReadResult, TxLintResult, TxExecResult, all output builders,
  exit_code_from_tx_output
- tx/execute.rs (1363 lines): TxState, CachedDoc, all execute_*_op
  functions, execute_operation dispatch, execute_and_collect
- tx/lifecycle.rs (461 lines): lifecycle steps, commit_changes,
  rollback_strict, RestoreFailGuard, execute_plan_direct

All crate::tx::* import paths continue to work via re-exports in
tx/mod.rs. Zero behavioral change.
