+++ src/tx/execute.rs
@@ -1,7 +1,6 @@
-#![cfg_attr(not(feature = "cli"), allow(dead_code, unused_imports))]
+use super::output::{TxExecResult, TxLintResult, TxReadResult, TxSearchMatch, TxSearchResult};
+use super::validate::op_label;
 use crate::cli::global::{EolMode, GlobalFlags};
-
-use crate::exit;
 use crate::ops::doc::{
     DocMutation, FileFormat, MutationResult, apply_doc_mutation, detect_format, parse_doc,
     serialize_value_preserving,
@@ -12,256 +11,16 @@ use crate::ops::md::{
 };
 use crate::ops::patch::{ApplyHunksOptions, ApplyHunksStatus, apply_patch_with_loader};
 use crate::ops::replace::{
-    ReplaceModeError, compile_replace_regex, replace_content, replace_whole_lines,
-    replacement_text, validate_replace_mode,
+    compile_replace_regex, replace_content, replace_whole_lines, replacement_text,
 };
-use crate::plan::{self, Operation, Plan};
-
-use crate::write::{WritePolicy, apply_policy, atomic_create_new, atomic_write};
-use anyhow::Context;
-#[cfg(feature = "cli")]
-use clap::Args;
+use crate::plan::{Operation, Plan};
+use crate::write::{WritePolicy, apply_policy};
 use globset::Glob;
 use ignore::WalkBuilder;
 use regex::RegexBuilder;
-use serde::{Deserialize, Serialize};
 use std::collections::hash_map::Entry;
 use std::collections::{HashMap, HashSet};
-
 use std::path::{Path, PathBuf};
-
-/// Structured report from `execute_plan` (and the `tx` command).
-///
-/// Library users can deserialize the JSON string returned by `execute_plan`
-/// into this type for typed access instead of string parsing.
-/// See #805 and the embedding docs.
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct TxOutput {
-    pub ok: bool,
-    pub status: String,
-    pub files_changed: usize,
-    pub files_created: usize,
-    pub files_deleted: usize,
-    pub changes: Vec<TxChange>,
-    #[serde(skip_serializing_if = "Vec::is_empty", default)]
-    pub reads: Vec<TxReadResult>,
-    #[serde(skip_serializing_if = "Vec::is_empty", default)]
-    pub searches: Vec<TxSearchResult>,
-    #[serde(skip_serializing_if = "Vec::is_empty", default)]
-    pub lints: Vec<TxLintResult>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub error_kind: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub error: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub backup_session: Option<String>,
-}
-
-/// A single file change in a plan/tx report.
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct TxChange {
-    pub path: String,
-    pub action: String,
-}
-
-/// A search match in the tx output.
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct TxSearchMatch {
-    pub line: usize,
-    pub column: usize,
-    pub text: String,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub context_before: Vec<String>,
-    #[serde(skip_serializing_if = "Vec::is_empty")]
-    pub context_after: Vec<String>,
-}
-
-/// A search result in the tx output.
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct TxSearchResult {
-    pub path: String,
-    pub pattern: String,
-    pub match_count: usize,
-    pub matches: Vec<TxSearchMatch>,
-}
-
-/// A file read result in the tx output.
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct TxReadResult {
-    pub path: String,
-    pub content: String,
-    pub start_line: usize,
-    pub end_line: usize,
-    pub total_lines: usize,
-}
-
-/// A lint result in the tx output.
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct TxLintResult {
-    pub path: String,
-    pub issue_count: usize,
-    pub issues: Vec<crate::ops::md::LintIssue>,
-}
-
-#[cfg(feature = "cli")]
-#[derive(Debug, Args)]
-#[non_exhaustive]
-#[command(after_help = "\
-EXAMPLES:
-  patchloom tx plan.json
-  patchloom tx plan.json --apply
-  patchloom tx plan.yaml --check --json")]
-pub struct TxArgs {
-    // ref:tx-mode:plan-stdin
-    /// Path to a plan file (JSON/YAML/TOML), or `-` for stdin.
-    pub plan: String,
-    // ref:tx-mode:plan-yaml
-    /// Plan format when reading from stdin (json, yaml, toml). Auto-detected from file extension otherwise.
-    #[arg(long)]
-    pub plan_format: Option<String>,
-    /// Disable strict rollback on format/validate failure.
-    #[arg(long)]
-    pub no_strict: bool,
-
-    #[command(flatten)]
-    pub write: crate::cli::global::WriteFlags,
-}
-
-const DEFAULT_LIFECYCLE_TIMEOUT_SECS: u64 = 60;
-use crate::exec;
-
-/// Short label for an operation, used in error messages.
-fn op_label(op: &Operation) -> &'static str {
-    match op {
-        Operation::Replace { .. } => "replace",
-        Operation::DocSet { .. } => "doc.set",
-        Operation::DocDelete { .. } => "doc.delete",
-        Operation::DocMerge { .. } => "doc.merge",
-        Operation::DocAppend { .. } => "doc.append",
-        Operation::DocPrepend { .. } => "doc.prepend",
-        Operation::DocUpdate { .. } => "doc.update",
-        Operation::DocMove { .. } => "doc.move",
-        Operation::DocEnsure { .. } => "doc.ensure",
-        Operation::DocDeleteWhere { .. } => "doc.delete_where",
-        Operation::MdReplaceSection { .. } => "md.replace_section",
-        Operation::MdInsertAfterHeading { .. } => "md.insert_after_heading",
-        Operation::MdInsertBeforeHeading { .. } => "md.insert_before_heading",
-        Operation::MdUpsertBullet { .. } => "md.upsert_bullet",
-        Operation::MdTableAppend { .. } => "md.table_append",
-        Operation::MdMoveSection { .. } => "md.move_section",
-        Operation::MdDedupeHeadings { .. } => "md.dedupe_headings",
-        Operation::TidyFix { .. } => "tidy.fix",
-        Operation::FileAppend { .. } => "file.append",
-        Operation::FileCreate { .. } => "file.create",
-        Operation::FileDelete { .. } => "file.delete",
-        Operation::FileRename { .. } => "file.rename",
-        Operation::PatchApply { .. } => "patch.apply",
-        Operation::Read { .. } => "read",
-        Operation::Search { .. } => "search",
-        Operation::MdLintAgents { .. } => "md.lint_agents",
-        #[cfg(feature = "ast")]
-        Operation::AstRename { .. } => "ast.rename",
-        #[cfg(feature = "ast")]
-        Operation::AstReplace { .. } => "ast.replace",
-    }
-}
-
-pub(crate) fn validate_operation(op: &Operation) -> anyhow::Result<()> {
-    match op {
-        Operation::Replace {
-            from,
-            to,
-            insert_before,
-            insert_after,
-            nth,
-            ..
-        } => {
-            if from.is_empty() {
-                anyhow::bail!("replace operation requires a non-empty search pattern");
-            }
-            if *nth == Some(0) {
-                anyhow::bail!("replace nth is 1-based; use 1 for the first occurrence");
-            }
-            match validate_replace_mode(
-                to.is_some(),
-                insert_before.is_some(),
-                insert_after.is_some(),
-            ) {
-                Ok(()) => Ok(()),
-                Err(ReplaceModeError::MissingMode) => {
-                    anyhow::bail!(
-                        "replace operation requires one of to, insert_before, or insert_after"
-                    )
-                }
-                Err(ReplaceModeError::BothInsertModes) => {
-                    anyhow::bail!("insert_before and insert_after cannot both be set")
-                }
-                Err(ReplaceModeError::ToWithInsert) => {
-                    anyhow::bail!("to cannot be combined with insert_before or insert_after")
-                }
-            }
-        }
-        // Exhaustive match ensures the compiler flags new variants that may
-        // need validation constraints.
-        Operation::DocSet { .. }
-        | Operation::DocDelete { .. }
-        | Operation::DocMerge { .. }
-        | Operation::DocAppend { .. }
-        | Operation::DocPrepend { .. }
-        | Operation::DocUpdate { .. }
-        | Operation::DocMove { .. }
-        | Operation::DocEnsure { .. }
-        | Operation::DocDeleteWhere { .. }
-        | Operation::MdReplaceSection { .. }
-        | Operation::MdInsertAfterHeading { .. }
-        | Operation::MdInsertBeforeHeading { .. }
-        | Operation::MdUpsertBullet { .. }
-        | Operation::MdTableAppend { .. }
-        | Operation::MdDedupeHeadings { .. }
-        | Operation::TidyFix { .. }
-        | Operation::FileAppend { .. }
-        | Operation::FileCreate { .. }
-        | Operation::FileDelete { .. }
-        | Operation::FileRename { .. }
-        | Operation::Read { .. }
-        | Operation::MdLintAgents { .. }
-        | Operation::PatchApply { .. } => Ok(()),
-        #[cfg(feature = "ast")]
-        Operation::AstRename { .. } | Operation::AstReplace { .. } => Ok(()),
-        Operation::MdMoveSection { before, after, .. } => {
-            if before.is_none() && after.is_none() {
-                anyhow::bail!("md.move_section requires either 'before' or 'after'");
-            }
-            if before.is_some() && after.is_some() {
-                anyhow::bail!("md.move_section: 'before' and 'after' cannot both be set");
-            }
-            Ok(())
-        }
-        Operation::Search {
-            invert_match,
-            multiline,
-            literal,
-            regex,
-            ..
-        } => {
-            if *invert_match && *multiline {
-                anyhow::bail!("search: invert_match and multiline cannot be combined");
-            }
-            if *literal && *regex {
-                anyhow::bail!("search: literal and regex cannot be combined");
-            }
-            Ok(())
-        }
-    }
-}
-
-pub(crate) fn validate_plan_operations(plan: &Plan) -> anyhow::Result<()> {
-    for op in &plan.operations {
-        validate_operation(op)?;
-    }
-
-    Ok(())
-}
 
 // ---------------------------------------------------------------------------
 // Pending file changes
@@ -1493,266 +1252,8 @@ fn build_write_policy(
 }
 
 // ---------------------------------------------------------------------------
-// Diff output helper
-// ---------------------------------------------------------------------------
-
-pub(crate) fn build_tx_output(
-    status: &'static str,
-    ok: bool,
-    changes: &[(PathBuf, String, String)],
-    deletions: &HashSet<PathBuf>,
-    existed_before: &HashSet<PathBuf>,
-    cwd: &Path,
-) -> TxOutput {
-    let mut tx_changes = Vec::new();
-    let mut created = 0usize;
-    let mut deleted_count = 0usize;
-    let mut modified = 0usize;
-
-    let display_path = |p: &Path| -> String {
-        crate::files::relative_display(p, cwd)
-            .to_string_lossy()
-            .into_owned()
-    };
-
-    for (path, _original, _) in changes {
-        let path_str = display_path(path);
-        if deletions.contains(path) {
-            tx_changes.push(TxChange {
-                path: path_str,
-                action: "deleted".to_string(),
-            });
-            deleted_count += 1;
-        } else if !existed_before.contains(path) {
-            tx_changes.push(TxChange {
-                path: path_str,
-                action: "created".to_string(),
-            });
-            created += 1;
-        } else {
-            tx_changes.push(TxChange {
-                path: path_str,
-                action: "modified".to_string(),
-            });
-            modified += 1;
-        }
-    }
-    // Deletions not captured in changes (empty files).
-    for path in deletions {
-        if !changes.iter().any(|(c, _, _)| c == path) {
-            tx_changes.push(TxChange {
-                path: display_path(path),
-                action: "deleted".to_string(),
-            });
-            deleted_count += 1;
-        }
-    }
-
-    TxOutput {
-        ok,
-        status: status.to_string(),
-        files_changed: modified,
-        files_created: created,
-        files_deleted: deleted_count,
-        changes: tx_changes,
-        reads: Vec::new(),
-        searches: Vec::new(),
-        lints: Vec::new(),
-        error_kind: None,
-        error: None,
-        backup_session: None,
-    }
-}
-
-pub(crate) fn build_full_tx_output(
-    status: &'static str,
-    result: &mut TxExecResult,
-    cwd: &Path,
-) -> TxOutput {
-    let mut output = build_tx_output(
-        status,
-        true,
-        &result.changes,
-        &result.deletions,
-        &result.existed_before,
-        cwd,
-    );
-    output.reads = std::mem::take(&mut result.tx_reads);
-    output.searches = std::mem::take(&mut result.tx_searches);
-    output.lints = std::mem::take(&mut result.tx_lints);
-    output
-}
-
-pub(crate) fn describe_exit_status(status: std::process::ExitStatus) -> String {
-    match status.code() {
-        Some(code) => format!("exit code {code}"),
-        None => "terminated by signal".to_string(),
-    }
-}
-
-pub(crate) fn describe_lifecycle_cwd(base_cwd: &Path, cwd: &Path) -> String {
-    if cwd == base_cwd {
-        ".".to_string()
-    } else {
-        crate::files::relative_display(cwd, base_cwd)
-            .display()
-            .to_string()
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Lifecycle helpers (extracted from run() for testability)
-// ---------------------------------------------------------------------------
-
-pub(crate) struct LifecycleError {
-    pub(crate) message: String,
-    pub(crate) kind: &'static str,
-}
-
-/// Format a lifecycle step failure message, appending stderr output if available.
-fn lifecycle_failure_msg(header: &str, stderr: &str) -> String {
-    let trimmed = stderr.trim();
-    if trimmed.is_empty() {
-        header.to_string()
-    } else {
-        format!("{header}: {trimmed}")
-    }
-}
-
-/// Shared runner for format and validation lifecycle steps.
-///
-/// `label` is used in error messages (e.g. "format step" or "required validation").
-/// `error_label` is the shorter variant for the error branch (e.g. "format step" or "validation").
-/// `kind` is the `LifecycleError::kind` string.
-/// `fail_on_error` controls whether non-required failures are fatal.
-fn run_lifecycle_steps(
-    steps: impl Iterator<Item = (String, Option<u64>, bool)>,
-    base_cwd: &Path,
-    cwd: &Path,
-    label: &str,
-    error_label: &str,
-    kind: &'static str,
-) -> Result<(), LifecycleError> {
-    let lifecycle_cwd = describe_lifecycle_cwd(base_cwd, cwd);
-    for (index, (cmd, timeout, required)) in steps.enumerate() {
-        let timeout_secs = timeout.unwrap_or(DEFAULT_LIFECYCLE_TIMEOUT_SECS);
-        let result = exec::run_with_timeout(&cmd, timeout_secs, cwd);
-        match result {
-            Ok(exec::ShellResult {
-                status,
-                stderr_head,
-            }) if !status.success() => {
-                let header = format!(
-                    "{label} failed (step {}, {}, cwd: {})",
-                    index + 1,
-                    describe_exit_status(status),
-                    lifecycle_cwd
-                );
-                let msg = lifecycle_failure_msg(&header, &stderr_head);
-                eprintln!("tx: {msg}");
-                if required {
-                    return Err(LifecycleError { message: msg, kind });
-                }
-            }
-            Err(e) => {
-                let msg = format!(
-                    "{error_label} error (step {}, cwd: {}): {e}",
-                    index + 1,
-                    lifecycle_cwd
-                );
-                eprintln!("tx: {msg}");
-                if required {
-                    return Err(LifecycleError { message: msg, kind });
-                }
-            }
-            _ => {}
-        }
-    }
-    Ok(())
-}
-
-pub(crate) fn run_format_steps(
-    steps: &[plan::FormatStep],
-    base_cwd: &Path,
-    cwd: &Path,
-) -> Result<(), LifecycleError> {
-    run_lifecycle_steps(
-        steps.iter().map(|s| (s.cmd.clone(), s.timeout, true)),
-        base_cwd,
-        cwd,
-        "format step",
-        "format step",
-        "format_failed",
-    )
-}
-
-pub(crate) fn run_validate_steps(
-    steps: &[plan::ValidationStep],
-    base_cwd: &Path,
-    cwd: &Path,
-) -> Result<(), LifecycleError> {
-    run_lifecycle_steps(
-        steps
-            .iter()
-            .map(|s| (s.cmd.clone(), s.timeout, s.required.unwrap_or(false))),
-        base_cwd,
-        cwd,
-        "required validation",
-        "validation",
-        "validation_failed",
-    )
-}
-
-pub(crate) fn rollback_strict(
-    changes: &[(PathBuf, String, String)],
-    pending: &HashMap<PathBuf, (String, String)>,
-    deletions: &HashSet<PathBuf>,
-    existed_before: &HashSet<PathBuf>,
-) {
-    let noop_policy = WritePolicy::default();
-    for (path, original, _) in changes {
-        if !existed_before.contains(path) && !deletions.contains(path) {
-            if let Err(e) = std::fs::remove_file(path) {
-                eprintln!("tx: rollback: failed to remove {}: {e}", path.display());
-            }
-        } else if let Err(e) = atomic_write(path, original, &noop_policy) {
-            eprintln!("tx: rollback: failed to restore {}: {e}", path.display());
-        }
-    }
-    for path in deletions {
-        if let Some((orig, _)) = pending.get(path)
-            && existed_before.contains(path)
-            && let Err(e) = atomic_write(path, orig, &noop_policy)
-        {
-            eprintln!(
-                "tx: rollback: failed to restore deleted {}: {e}",
-                path.display()
-            );
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
 // Shared execution core
 // ---------------------------------------------------------------------------
-
-/// Intermediate result from executing all operations in a plan and applying
-/// write policy. Contains everything needed for callers to decide on output
-/// mode, commit changes, and run lifecycle steps.
-pub(crate) struct TxExecResult {
-    pub(crate) changes: Vec<(PathBuf, String, String)>,
-    pub(crate) deletions: HashSet<PathBuf>,
-    pub(crate) existed_before: HashSet<PathBuf>,
-    /// Original pending map, retained for `rollback_strict`.
-    pub(crate) pending: HashMap<PathBuf, (String, String)>,
-    pub(crate) tx_reads: Vec<TxReadResult>,
-    pub(crate) tx_searches: Vec<TxSearchResult>,
-    pub(crate) tx_lints: Vec<TxLintResult>,
-    pub(crate) no_effective_changes: bool,
-    pub(crate) replace_no_matches: bool,
-    /// "Did you mean?" hints when a replace found zero matches.
-    pub(crate) replace_hint: Option<String>,
-}
 
 /// Execute all plan operations in memory, apply write policy, and return
 /// the collected results without touching the filesystem.
@@ -1860,372 +1361,3 @@ pub(crate) fn execute_and_collect(
         replace_hint,
     })
 }
-
-/// Run format and validation lifecycle steps. Returns `None` on success.
-pub(crate) fn run_lifecycle(plan: &Plan, base_cwd: &Path, cwd: &Path) -> Option<LifecycleError> {
-    plan.format
-        .as_deref()
-        .map(|steps| run_format_steps(steps, base_cwd, cwd))
-        .unwrap_or(Ok(()))
-        .err()
-        .or_else(|| {
-            plan.validate
-                .as_deref()
-                .and_then(|steps| run_validate_steps(steps, base_cwd, cwd).err())
-        })
-}
-
-pub(crate) fn resolve_plan_cwd(base_cwd: &Path, plan_cwd: Option<&str>) -> PathBuf {
-    match plan_cwd {
-        Some(plan_cwd) => {
-            let path = Path::new(plan_cwd);
-            if path.is_absolute() {
-                path.to_path_buf()
-            } else {
-                base_cwd.join(path)
-            }
-        }
-        None => base_cwd.to_path_buf(),
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Direct execution (MCP / in-process callers)
-// ---------------------------------------------------------------------------
-
-/// Failure while committing staged changes to disk.
-#[derive(Debug)]
-pub struct CommitError {
-    pub message: String,
-    pub rollback_ok: bool,
-    pub backup_session: Option<String>,
-}
-
-impl std::fmt::Display for CommitError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.message)
-    }
-}
-
-impl std::error::Error for CommitError {}
-
-fn commit_error(message: impl Into<String>) -> CommitError {
-    CommitError {
-        message: message.into(),
-        rollback_ok: true,
-        backup_session: None,
-    }
-}
-
-thread_local! {
-    /// Per-thread test hook; never set in production.
-    static FORCE_RESTORE_FAIL: std::sync::atomic::AtomicBool =
-        const { std::sync::atomic::AtomicBool::new(false) };
-}
-
-/// RAII guard that forces restore failure on the current thread only.
-#[doc(hidden)]
-pub struct RestoreFailGuard;
-
-impl RestoreFailGuard {
-    pub fn engage() -> Self {
-        FORCE_RESTORE_FAIL.with(|flag| flag.store(true, std::sync::atomic::Ordering::SeqCst));
-        Self
-    }
-}
-
-impl Drop for RestoreFailGuard {
-    fn drop(&mut self) {
-        FORCE_RESTORE_FAIL.with(|flag| flag.store(false, std::sync::atomic::Ordering::SeqCst));
-    }
-}
-
-/// Restore files from a backup session after a failed commit. Returns `true`
-/// when restore completed successfully.
-pub(crate) fn restore_after_failed_commit(cwd: &Path, timestamp: &str) -> bool {
-    if FORCE_RESTORE_FAIL.with(|flag| flag.load(std::sync::atomic::Ordering::SeqCst)) {
-        return false;
-    }
-    crate::backup::restore_session(cwd, timestamp).is_ok()
-}
-
-/// Apply pending changes to disk: backup originals, write modified files,
-/// delete removed files, finalize backup session.
-///
-/// If any write fails, restores all already-written files from the backup
-/// session before returning [`CommitError`].
-pub(crate) fn commit_changes(
-    changes: &[(PathBuf, String, String)],
-    deletions: &HashSet<PathBuf>,
-    existed_before: &HashSet<PathBuf>,
-    cwd: &Path,
-) -> Result<(), CommitError> {
-    let mut backup = crate::backup::BackupSession::new(cwd)
-        .map_err(|e| commit_error(format!("starting backup session: {e}")))?;
-    for (path, _, _) in changes {
-        if deletions.contains(path) {
-            backup
-                .save_before_delete(path)
-                .map_err(|e| commit_error(format!("backing up {}: {e}", path.display())))?;
-        } else {
-            backup
-                .save_before_write(path)
-                .map_err(|e| commit_error(format!("backing up {}: {e}", path.display())))?;
-        }
-    }
-    for path in deletions {
-        backup
-            .save_before_delete(path)
-            .map_err(|e| commit_error(format!("backing up {}: {e}", path.display())))?;
-    }
-
-    // Finalize before writes so undo can recover from a mid-commit failure.
-    let backup_session = backup
-        .finalize()
-        .map_err(|e| commit_error(format!("finalizing backup session: {e}")))?;
-
-    let noop_policy = WritePolicy::default();
-    let write_result = (|| -> anyhow::Result<()> {
-        for (path, _, new_content) in changes {
-            if deletions.contains(path) {
-                std::fs::remove_file(path)
-                    .with_context(|| format!("deleting {}", path.display()))?;
-            } else {
-                if let Some(parent) = path.parent()
-                    && !parent.as_os_str().is_empty()
-                    && !parent.exists()
-                {
-                    std::fs::create_dir_all(parent)
-                        .with_context(|| format!("creating directory {}", parent.display()))?;
-                }
-                if !existed_before.contains(path) {
-                    atomic_create_new(path, new_content, &noop_policy)?;
-                } else {
-                    atomic_write(path, new_content, &noop_policy)?;
-                }
-            }
-        }
-        for path in deletions {
-            if path.exists() {
-                std::fs::remove_file(path)
-                    .with_context(|| format!("deleting {}", path.display()))?;
-            }
-        }
-        Ok(())
-    })();
-
-    if let Err(e) = write_result {
-        let rollback_ok = if let Some(ref ts) = backup_session {
-            restore_after_failed_commit(cwd, ts)
-        } else {
-            true
-        };
-        return Err(CommitError {
-            message: e.to_string(),
-            rollback_ok,
-            backup_session,
-        });
-    }
-
-    Ok(())
-}
-
-pub(crate) fn format_error_with_backup_hint(error: &str, backup_session: Option<&str>) -> String {
-    match backup_session {
-        Some(ts) => format!("{error} (backup session {ts}; run `patchloom undo` to restore)"),
-        None => error.to_string(),
-    }
-}
-
-pub(crate) fn build_error_output(
-    error_kind: &'static str,
-    legacy_error_prefix: &str,
-    error: &str,
-    backup_session: Option<&str>,
-) -> TxOutput {
-    TxOutput {
-        ok: false,
-        status: "error".to_string(),
-        files_changed: 0,
-        files_created: 0,
-        files_deleted: 0,
-        changes: Vec::new(),
-        reads: Vec::new(),
-        searches: Vec::new(),
-        lints: Vec::new(),
-        error_kind: Some(error_kind.to_string()),
-        error: Some(format!(
-            "{legacy_error_prefix}: {}",
-            format_error_with_backup_hint(error, backup_session)
-        )),
-        backup_session: backup_session.map(str::to_string),
-    }
-}
-
-fn config_tx_strict(cwd: &Path) -> Option<bool> {
-    crate::config::find_and_load(cwd)
-        .map(|(config, _)| config.tx.strict)
-        .unwrap_or(None)
-}
-
-/// Execute a parsed [`Plan`] directly and return the structured `TxOutput` (PlanReport).
-/// Does **not** write to stdout or stderr.
-///
-/// This is the in-process equivalent used by the library API (`api::execute_plan`)
-/// and (via serialization) by the CLI tx command and MCP.
-/// Library users get a typed `PlanReport` directly (addresses #811).
-#[allow(clippy::result_large_err)]
-pub(crate) fn validate_and_prepare_plan(
-    plan: &Plan,
-    cwd: &Path,
-    no_strict: bool,
-) -> Result<(PathBuf, bool, GlobalFlags), TxOutput> {
-    if plan.version != crate::plan::SCHEMA_VERSION {
-        let msg = format!(
-            "unsupported plan version '{}' (this build supports version {})",
-            plan.version,
-            crate::plan::SCHEMA_VERSION
-        );
-        return Err(build_error_output("parse_error", "parse_error", &msg, None));
-    }
-    if let Err(e) = validate_plan_operations(plan) {
-        return Err(build_error_output(
-            "parse_error",
-            "parse_error",
-            &e.to_string(),
-            None,
-        ));
-    }
-
-    let effective_cwd = resolve_plan_cwd(cwd, plan.cwd.as_deref());
-    let config_strict = config_tx_strict(&effective_cwd);
-    let strict = plan::effective_strict(plan.strict, config_strict, no_strict);
-
-    let mut global = GlobalFlags::with_cwd(&effective_cwd);
-    if let Some((config, _)) = crate::config::find_and_load(&effective_cwd) {
-        crate::config::apply_config(&mut global, &config);
-    }
-
-    Ok((effective_cwd, strict, global))
-}
-
-#[allow(clippy::result_large_err)]
-pub fn execute_plan_direct(
-    plan: Plan,
-    cwd: &Path,
-    guard: Option<&crate::containment::PathGuard>,
-) -> anyhow::Result<TxOutput> {
-    crate::verbose!(
-        "tx: direct plan execution ({} ops, cwd={}, guard={})",
-        plan.operations.len(),
-        cwd.display(),
-        guard.is_some()
-    );
-
-    let (effective_cwd, strict, global) = match validate_and_prepare_plan(&plan, cwd, false) {
-        Ok(v) => v,
-        Err(output) => return Ok(output),
-    };
-
-    // PathGuard enforcement for library callers of execute_plan (addresses #755).
-    // Upfront check on declared paths using shared helper; dynamic (globs
-    // patterns, patch embedded paths) are best-effort or handled by loaders.
-    if let Some(g) = guard {
-        for op in &plan.operations {
-            for p in crate::plan::declared_paths(op) {
-                g.check_path(p)
-                    .map_err(|e| anyhow::anyhow!("path rejected by workspace guard: {}", e))?;
-            }
-        }
-    }
-
-    // Execute operations and collect changes in memory.
-    let mut result = match execute_and_collect(&plan, &effective_cwd, &global, true, true) {
-        Ok(r) => r,
-        Err(e) => {
-            return Ok(build_error_output(
-                "operation_failed",
-                "operation_failed",
-                &e.to_string(),
-                None,
-            ));
-        }
-    };
-
-    if result.replace_no_matches {
-        let output = build_error_output(
-            "no_matches",
-            "no_matches",
-            result.replace_hint.as_deref().unwrap_or(""),
-            None,
-        );
-        return Ok(output);
-    }
-
-    if result.no_effective_changes {
-        let output = build_full_tx_output("success", &mut result, &effective_cwd);
-        return Ok(output);
-    }
-
-    // Apply: back up originals, write files.
-    if let Err(err) = commit_changes(
-        &result.changes,
-        &result.deletions,
-        &result.existed_before,
-        &effective_cwd,
-    ) {
-        let error_kind = if err.rollback_ok {
-            "rollback"
-        } else {
-            "rollback_failed"
-        };
-        let output = build_error_output(
-            error_kind,
-            error_kind,
-            &err.message,
-            err.backup_session.as_deref(),
-        );
-        return Ok(output);
-    }
-
-    // Run format steps, then validation steps.
-    if let Some(err) = run_lifecycle(&plan, cwd, &effective_cwd) {
-        if strict {
-            rollback_strict(
-                &result.changes,
-                &result.pending,
-                &result.deletions,
-                &result.existed_before,
-            );
-            let msg = format!("strict mode -- all changes reverted ({})", err.message);
-            return Ok(build_error_output(err.kind, "rollback", &msg, None));
-        }
-        return Ok(build_error_output(err.kind, err.kind, &err.message, None));
-    }
-
-    let output = build_full_tx_output("success", &mut result, &effective_cwd);
-    Ok(output)
-}
-
-/// Map a `TxOutput` (PlanReport) to the traditional exit code for CLI/MCP compat.
-pub fn exit_code_from_tx_output(report: &TxOutput) -> u8 {
-    if report.ok {
-        if report.status == "no_matches" {
-            exit::NO_MATCHES
-        } else {
-            exit::SUCCESS
-        }
-    } else {
-        match report.error_kind.as_deref() {
-            Some("no_matches") => exit::NO_MATCHES,
-            Some("parse_error") => exit::PARSE_ERROR,
-            Some("rollback") => exit::ROLLBACK,
-            Some("rollback_failed") => exit::FAILURE, // preserve historical CLI exit for this case
-            Some("validation_failed") | Some("format_failed") => exit::VALIDATION_FAILED,
-            Some("operation_failed") => exit::OPERATION_FAILED,
-            _ => exit::FAILURE,
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
